requirements.txt
backports.tarfile==1.2.0
boto3==1.34.144
botocore==1.34.144
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.3
cryptography==42.0.8
docutils==0.21.2
ghp-import==2.1.0
griffe==0.47.0
idna==3.7
importlib_metadata==7.1.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
keyring==25.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-autorefs==1.0.1
mkdocs-get-deps==0.2.0
mkdocstrings==0.25.1
mkdocstrings-python==1.10.4
more-itertools==10.3.0
nh3==0.2.17
packaging==24.1
pathspec==0.12.1
pkginfo==1.11.1
platformdirs==4.2.2
pycparser==2.22
Pygments==2.18.0
pymdown-extensions==10.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
readme_renderer==43.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
s3transfer==0.10.2
SecretStorage==3.3.3
six==1.16.0
twine==5.1.0
urllib3==2.2.2
watchdog==4.0.1
zipp==3.19.2
